+++ spec/subdir/dom-test.js
@@ -5,8 +5,7 @@ atom.document.body.className += " nah";
 describe("DOM", () => {
 	
 	it("shows what a failed class assertion looks like", () => {
-		const body = atom.document.body;
-		const html = atom.document.documentElement;
+		const {body, documentElement: html} = atom.document;
 		
 		html.classList.add("root");
 		expect([body, html]).to.have.classes("platform-darwin", ["nah"]);